+++ src/entities/trouble/ui/PostGuideMd.tsx
