+++ src/layouts/Header/Header.tsx
@@ -6,8 +6,8 @@ import { FaUserCircle } from "react-icons/fa";
 import { MdSearch } from "react-icons/md";
 import { useSearchStore } from "@/store/useSearchStore";
 import useClickOutside from "@/hooks/useClickOutside";
-import UserMenuDropdown from "../Menu/UserMenuDropdown";
-import NotificationModal from "../Modal/NotificationModal";
+import UserMenuDropdown from "../../shared/ui/Menu/UserMenuDropdown";
+import NotificationModal from "../../shared/ui/Modal/NotificationModal";
 import { PATH } from "@/shared/config/paths";
 import logo from "@/assets/icons/logo.svg";
 import { useViewerId } from "@/store/auth";
