+++ src/pages/Onboarding/IntroLandingPage.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { PATH } from "@/shared/config/paths";
-import HeaderLogoOnly from "@/components/Header/HeaderLogoOnly";
+import HeaderLogoOnly from "@/layouts/Header/HeaderLogoOnly";
 import HeroLaptop from "@/assets/images/hero-laptop.png";
 import GuideShot from "@/assets/images/guide-shot.png";
 import GuideChecklist from "@/assets/images/guide-checklist.png";
