+++ src/shared/ui/Editor/EditorBlock.tsx
