+++ src/shared/ui/Menu/SummaryTypeDropdown.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import useClickOutside from "../../hooks/useClickOutside";
+import useClickOutside from "../../../hooks/useClickOutside";
 import dropdownIcon from "@/assets/icons/dropdown_icon.svg";
 import type { ProjectTroubleSummaryType } from "@/entities/trouble/model";
