+++ src/shared/ui/Modal/ConfirmDeleteModal.tsx
