+++ src/shared/ui/Modal/WithdrawCompleteModal.tsx
